Use uint64_t for /proc/stat cpu parsed type (#29)

+++ system_metrics_collector/src/system_metrics_collector/proc_cpu_data.cpp
@@ -22,13 +22,13 @@ namespace system_metrics_collector
 
 /*static*/ constexpr const char ProcCpuData::EMPTY_LABEL[];
 
-size_t ProcCpuData::getIdleTime() const
+uint64_t ProcCpuData::getIdleTime() const
 {
   return times[static_cast<int>(ProcCpuStates::kIdle)] +
          times[static_cast<int>(ProcCpuStates::kIOWait)];
 }
 
-size_t ProcCpuData::getActiveTime() const
+uint64_t ProcCpuData::getActiveTime() const
 {
   return times[static_cast<int>(ProcCpuStates::kUser)] +
          times[static_cast<int>(ProcCpuStates::kNice)] +
